Update OAS and Auth helpers
