+++ test/library/encryption/assertion.go
@@ -23,6 +23,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	configv1 "github.com/openshift/api/config/v1"
+	routev1 "github.com/openshift/api/route/v1"
 )
 
 var protoEncodingPrefix = []byte{0x6b, 0x38, 0x73, 0x00}
@@ -231,5 +232,90 @@ func assertConfigMaps(t testing.TB, etcdClient EtcdClient, expectedMode string) 
 	t.Logf("Checking if all ConfigMaps where encrypted/decrypted for %q mode", expectedMode)
 	totalConfigMaps, err := VerifyResources(t, etcdClient, "/kubernetes.io/configmaps/", expectedMode, false)
 	t.Logf("Verified %d ConfigMaps", totalConfigMaps)
+	require.NoError(t, err)
+}
+
+var WellKnownAuthTargetGRs = []schema.GroupResource{
+	{Group: "oauth.openshift.io", Resource: "oauthaccesstokens"},
+	{Group: "oauth.openshift.io", Resource: "oauthauthorizetokens"},
+}
+
+func AssertWellKnownTokenOfLifeEncrypted(t testing.TB, clientSet ClientSet, _ runtime.Object) {
+	t.Helper()
+	rawTokenValue := GetRawWellKnownTokenOfLife(t, clientSet)
+	marker := "I have no special talents. I am only passionately curious"
+	if strings.Contains(rawTokenValue, marker) {
+		t.Errorf("access token not encrypted, etcd value contains refresh token marker in plain text")
+	}
+}
+
+func AssertWellKnownTokenOfLifeNotEncrypted(t testing.TB, clientSet ClientSet, _ runtime.Object) {
+	t.Helper()
+	rawTokenValue := GetRawWellKnownTokenOfLife(t, clientSet)
+	marker := "I have no special talents. I am only passionately curious"
+	if !strings.Contains(rawTokenValue, marker) {
+		t.Errorf("access token not decrypted, etcd value does not contain refresh token marker in plain text")
+	}
+}
+
+func AssertWellKnownTokens(t testing.TB, clientSet ClientSet, expectedMode configv1.EncryptionType, namespace, labelSelector string) {
+	t.Helper()
+	assertWellKnownAccessTokens(t, clientSet.Etcd, string(expectedMode))
+	assertWellKnownAuthTokens(t, clientSet.Etcd, string(expectedMode))
+	AssertLastMigratedKey(t, clientSet.Kube, WellKnownAuthTargetGRs, namespace, labelSelector)
+}
+
+func assertWellKnownAccessTokens(t testing.TB, etcdClient EtcdClient, expectedMode string) {
+	t.Logf("Checking if all OauthAccessTokens where encrypted/decrypted for %q mode", expectedMode)
+	totalAccessTokens, err := VerifyResources(t, etcdClient, "/openshift.io/oauth/accesstokens/", expectedMode, true)
+	t.Logf("Verified %d OauthAccessTokens", totalAccessTokens)
+	require.NoError(t, err)
+}
+
+func assertWellKnownAuthTokens(t testing.TB, etcdClient EtcdClient, expectedMode string) {
+	t.Logf("Checking if all OAuthAuthorizeTokens where encrypted/decrypted for %q mode", expectedMode)
+	totalAuthTokens, err := VerifyResources(t, etcdClient, "/openshift.io/oauth/authorizetokens/", expectedMode, true)
+	t.Logf("Verified %d OAuthAuthorizeTokens", totalAuthTokens)
+	require.NoError(t, err)
+}
+
+var WellKnownOASTargetGRs = []schema.GroupResource{
+	{Group: "route.openshift.io", Resource: "routes"},
+}
+
+func AssertWellKnownRouteOfLifeEncrypted(t testing.TB, clientSet ClientSet, resource runtime.Object) {
+	t.Helper()
+	routeOfLife, ok := resource.(*routev1.Route)
+	if !ok {
+		t.Fatalf("expected *routev1.Route, got %T", resource)
+	}
+	rawRouteValue := GetRawWellKnownRouteOfLife(t, clientSet, routeOfLife.Namespace)
+	if strings.Contains(rawRouteValue, routeOfLife.Spec.To.Name) {
+		t.Errorf("route not encrypted, etcd value contains target name %q in plain text", routeOfLife.Spec.To.Name)
+	}
+}
+
+func AssertWellKnownRouteOfLifeNotEncrypted(t testing.TB, clientSet ClientSet, resource runtime.Object) {
+	t.Helper()
+	routeOfLife, ok := resource.(*routev1.Route)
+	if !ok {
+		t.Fatalf("expected *routev1.Route, got %T", resource)
+	}
+	rawRouteValue := GetRawWellKnownRouteOfLife(t, clientSet, routeOfLife.Namespace)
+	if !strings.Contains(rawRouteValue, routeOfLife.Spec.To.Name) {
+		t.Errorf("route not decrypted, etcd value does not contain target name %q in plain text", routeOfLife.Spec.To.Name)
+	}
+}
+
+func AssertWellKnownRoutes(t testing.TB, clientSet ClientSet, expectedMode configv1.EncryptionType, namespace, labelSelector string) {
+	t.Helper()
+	assertWellKnownRoutes(t, clientSet.Etcd, string(expectedMode))
+	AssertLastMigratedKey(t, clientSet.Kube, WellKnownOASTargetGRs, namespace, labelSelector)
+}
+
+func assertWellKnownRoutes(t testing.TB, etcdClient EtcdClient, expectedMode string) {
+	t.Logf("Checking if all Routes where encrypted/decrypted for %q mode", expectedMode)
+	totalRoutes, err := VerifyResources(t, etcdClient, "/openshift.io/routes/", expectedMode, false)
+	t.Logf("Verified %d Routes", totalRoutes)
 	require.NoError(t, err)
 }
