+++ test/library/encryption/helpers.go
@@ -28,7 +28,11 @@ import (
 	configv1 "github.com/openshift/api/config/v1"
 	configv1client "github.com/openshift/client-go/config/clientset/versioned/typed/config/v1"
 
+	oauthapiv1 "github.com/openshift/api/oauth/v1"
+	routev1 "github.com/openshift/api/route/v1"
 	"github.com/openshift/library-go/test/library"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
 var (
@@ -562,4 +566,120 @@ func inParallel(steps ...testStep) testStep {
 			wg.Wait()
 		},
 	}
+}
+
+var wellKnownRouteGVR = schema.GroupVersionResource{Group: "route.openshift.io", Version: "v1", Resource: "routes"}
+
+func CreateAndStoreWellKnownRouteOfLife(ctx context.Context, t testing.TB, cs ClientSet, ns string) runtime.Object {
+	t.Helper()
+	t.Logf("Creating %q in %q namespace", "route-of-life", ns)
+
+	route := WellKnownRouteOfLife(t, ns)
+	obj, err := runtime.DefaultUnstructuredConverter.ToUnstructured(route)
+	require.NoError(t, err)
+
+	created, err := cs.DynamicClient.Resource(wellKnownRouteGVR).Namespace(ns).Create(ctx, &unstructured.Unstructured{Object: obj}, metav1.CreateOptions{})
+	require.NoError(t, err)
+
+	var result routev1.Route
+	err = runtime.DefaultUnstructuredConverter.FromUnstructured(created.Object, &result)
+	require.NoError(t, err)
+	return &result
+}
+
+func WellKnownRouteOfLife(_ testing.TB, ns string) runtime.Object {
+	return &routev1.Route{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "route.openshift.io/v1",
+			Kind:       "Route",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "route-of-life",
+			Namespace: ns,
+		},
+		Spec: routev1.RouteSpec{
+			Host: "devcluster.openshift.io",
+			Port: &routev1.RoutePort{
+				TargetPort: intstr.FromInt(2014),
+			},
+			To: routev1.RouteTargetReference{
+				Name: "dummyroute",
+			},
+		},
+	}
+}
+
+func GetRawWellKnownRouteOfLife(t testing.TB, clientSet ClientSet, ns string) string {
+	t.Helper()
+	timeout, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
+	defer cancel()
+
+	routeOfLifeKey := fmt.Sprintf("/openshift.io/routes/%s/%s", ns, "route-of-life")
+	resp, err := clientSet.Etcd.Get(timeout, routeOfLifeKey)
+	require.NoError(t, err)
+	require.Len(t, resp.Kvs, 1, "expected exactly one key from etcd for route-of-life")
+
+	return string(resp.Kvs[0].Value)
+}
+
+var wellKnownOAuthAccessTokenGVR = schema.GroupVersionResource{Group: "oauth.openshift.io", Version: "v1", Resource: "oauthaccesstokens"}
+
+const wellKnownTokenOfLifeName = "sha256~token-aaaaaaaa-of-aaaaaaaa-life-aaaaaaaa"
+
+func CreateAndStoreWellKnownTokenOfLife(ctx context.Context, t testing.TB, cs ClientSet) runtime.Object {
+	t.Helper()
+	tokens := cs.DynamicClient.Resource(wellKnownOAuthAccessTokenGVR)
+
+	oldToken, err := tokens.Get(ctx, wellKnownTokenOfLifeName, metav1.GetOptions{})
+	if err != nil && !errors.IsNotFound(err) {
+		t.Fatalf("Failed to check if the token already exists: %v", err)
+	}
+	if oldToken != nil && len(oldToken.GetName()) > 0 {
+		t.Log("The access token already exists, removing it first")
+		require.NoError(t, tokens.Delete(ctx, oldToken.GetName(), metav1.DeleteOptions{}))
+	}
+
+	t.Logf("Creating %q at cluster scope level", wellKnownTokenOfLifeName)
+	token := WellKnownTokenOfLife(t, "")
+	obj, err := runtime.DefaultUnstructuredConverter.ToUnstructured(token)
+	require.NoError(t, err)
+
+	created, err := tokens.Create(ctx, &unstructured.Unstructured{Object: obj}, metav1.CreateOptions{})
+	require.NoError(t, err)
+
+	var result oauthapiv1.OAuthAccessToken
+	err = runtime.DefaultUnstructuredConverter.FromUnstructured(created.Object, &result)
+	require.NoError(t, err)
+	return &result
+}
+
+func WellKnownTokenOfLife(_ testing.TB, _ string) runtime.Object {
+	return &oauthapiv1.OAuthAccessToken{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "oauth.openshift.io/v1",
+			Kind:       "OAuthAccessToken",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: wellKnownTokenOfLifeName,
+		},
+		RefreshToken: "I have no special talents. I am only passionately curious",
+		UserName:     "kube:admin",
+		Scopes:       []string{"user:full"},
+		RedirectURI:  "redirect.me.to.token.of.life",
+		ClientName:   "console",
+		UserUID:      "non-existing-user-id",
+	}
+}
+
+func GetRawWellKnownTokenOfLife(t testing.TB, clientSet ClientSet) string {
+	t.Helper()
+	timeout, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
+	defer cancel()
+
+	tokenOfLifeKey := fmt.Sprintf("/openshift.io/oauth/accesstokens/%s", wellKnownTokenOfLifeName)
+	resp, err := clientSet.Etcd.Get(timeout, tokenOfLifeKey)
+	require.NoError(t, err)
+	require.Len(t, resp.Kvs, 1, "expected exactly one key from etcd for token-of-life")
+
+	return string(resp.Kvs[0].Value)
 }
